move/rename Subprocess variable to dtest/testprocess package
